:building_construction: moved eject functions to it proper file
