+++ Network/Makefile_Network.mk
@@ -32,6 +32,7 @@ SRC		= 		main.c								\
 				Game/object_manipulation.c			\
 				Game/incantation.c					\
 				Game/game_tick.c					\
+				Game/eject_function.c				\
 
 TRUE_SRC 	= 	$(patsubst %,src/%, $(SRC))
